+++ tests/EnumValueTest.php
@@ -5,6 +5,7 @@ namespace BenSampo\Enum\Tests;
 use PHPUnit\Framework\TestCase;
 use BenSampo\Enum\Rules\EnumValue;
 use BenSampo\Enum\Tests\Enums\UserType;
+use BenSampo\Enum\Tests\Enums\SuperPowers;
 use BenSampo\Enum\Tests\Enums\StringValues;
 
 class EnumValueTest extends TestCase
@@ -27,6 +28,47 @@ class EnumValueTest extends TestCase
         $this->assertFalse($fails1);
         $this->assertFalse($fails2);
         $this->assertFalse($fails3);
+    }
+
+    public function test_flagged_enum_passes_with_no_flags_set()
+    {
+        $passed = (new EnumValue(SuperPowers::class))->passes('', 0);
+
+        $this->assertTrue($passed);
+    }
+
+    public function test_flagged_enum_passes_with_single_flag_set()
+    {
+        $passed = (new EnumValue(SuperPowers::class))->passes('', SuperPowers::Flight);
+
+        $this->assertTrue($passed);
+    }
+
+    public function test_flagged_enum_passes_with_multiple_flags_set()
+    {
+        $passed = (new EnumValue(SuperPowers::class))->passes('', SuperPowers::Superman);
+
+        $this->assertTrue($passed);
+    }
+
+    public function test_flagged_enum_passes_with_all_flags_set()
+    {
+        $allFlags = array_reduce(SuperPowers::getValues(), function (int $carry, int $powerValue) {
+            return $carry | $powerValue;
+        }, 0);
+        $passed = (new EnumValue(SuperPowers::class))->passes('', $allFlags);
+
+        $this->assertTrue($passed);
+    }
+
+    public function test_flagged_enum_fails_with_invalid_flag_set()
+    {
+        $allFlagsSet = array_reduce(SuperPowers::getValues(), function ($carry, $value) {
+            return $carry | $value;
+        }, 0);
+        $passed = (new EnumValue(SuperPowers::class))->passes('', $allFlagsSet + 1);
+
+        $this->assertFalse($passed);
     }
 
     public function test_can_turn_off_strict_type_checking()